PyServer/requirements.txt
protobuf==3.20.0
grpcio==1.53.0
Pillow==9.5.0
torch==1.13.1
torchvision==0.14.1
